fix bugs after merge
