--- src/main/java/com/softserve/academy/event/repository/SurveyReporitory.java
@@ -1,6 +1,0 @@
-package com.softserve.academy.event.repository;
-
-import com.softserve.academy.event.entity.Survey;
-
-public interface SurveyReporitory extends BasicRepository<Survey, Long> {
-}
