+++ src/main/java/com/softserve/academy/event/repository/SurveyRepository.java
@@ -1,18 +1,17 @@
-package com.softserve.academy.event.service.db;
+package com.softserve.academy.event.repository;
 
-import com.softserve.academy.event.dto.SimpleSurveyDTO;
 import com.softserve.academy.event.entity.Survey;
 import com.softserve.academy.event.util.Page;
 import com.softserve.academy.event.util.Pageable;
 import org.springframework.http.HttpStatus;
 
-public interface SurveyService extends BasicService<Survey, Long> {
+public interface SurveyRepository extends BasicRepository<Survey, Long> {
 
-    Page<SimpleSurveyDTO> findAll(Pageable pageable);
+    Page<Survey> findAll(Pageable pageable);
 
     HttpStatus updateTitle(Long id, String title);
 
-    SimpleSurveyDTO duplicateSurvey(Long id);
+    Survey duplicateSurvey(Long id);
 
     String setTitleForSurvey(Long id, String title);
