+++ src/main/java/com/softserve/academy/event/service/db/BasicService.java
@@ -10,5 +10,6 @@ public interface BasicService<T extends Serializable, I extends Serializable> {
     List<T> findAll();
     T save(T entity);
     T update(T object);
-    void deleteById(I id);
+    void delete(T entity);
+    void detach(T entity);
 }
